yuzu/applets/software_keyboard: Resolve sign mismatch comparison
Qt uses a signed value to represent container sizes, so this was causing
a sign mismatch warning.

+++ src/yuzu/applets/software_keyboard.cpp
@@ -18,7 +18,7 @@ QtSoftwareKeyboardValidator::QtSoftwareKeyboardValidator(
     : parameters(std::move(parameters)) {}
 
 QValidator::State QtSoftwareKeyboardValidator::validate(QString& input, int& pos) const {
-    if (input.size() > parameters.max_length) {
+    if (input.size() > static_cast<s64>(parameters.max_length)) {
         return Invalid;
     }
     if (parameters.disable_space && input.contains(QLatin1Char{' '})) {
